+++ ledger/ledger-security-evidence/tool/src/scala/com/daml/security/evidence/generator/ScalaTestGeneratorSupport.scala
@@ -1,0 +1,47 @@
+// Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.security.evidence.generator
+
+import cats.syntax.either._
+import cats.syntax.functor._
+import cats.syntax.traverse._
+import cats.syntax.functorFilter._
+import io.circe.parser.decode
+import com.daml.security.evidence.tag.EvidenceTag
+import org.scalatest.Suite
+import com.daml.security.evidence.scalatest.JsonCodec._
+import org.scalatest.daml.ScalaTestAdapter
+import scala.reflect.ClassTag
+
+object ScalaTestGeneratorSupport {
+
+  private def testNameWithTags(tags: Map[String, Set[String]]): List[(String, List[EvidenceTag])] =
+    tags.fmap { tagNames =>
+      tagNames.toList
+        .filter(_.startsWith("{")) // Check if we have a JSON encoded tag
+        .traverse(decode[com.daml.security.evidence.tag.EvidenceTag])
+        .valueOr(err => sys.error(s"Failed to parse JSON tag: $err"))
+    }.toList
+
+  private def isIgnored(suite: Suite, testName: String): Boolean =
+    suite.tags.getOrElse(testName, Set()).contains(ScalaTestAdapter.IgnoreTagName)
+
+  def testEntries[TT: ClassTag, TS: ClassTag, TE](
+      suites: List[Suite],
+      testEntry: (String, String, TT, Boolean, Option[TS]) => TE,
+  ): List[TE] = {
+    suites.flatMap { suite =>
+      val testSuite = suite match {
+        case testSuite: TS => Some(testSuite)
+        case _ => None
+      }
+
+      testNameWithTags(suite.tags).mapFilter { case (testName, testTags) =>
+        testTags.collectFirst { case testTag: TT =>
+          testEntry(suite.suiteName, testName, testTag, isIgnored(suite, testName), testSuite)
+        }
+      }
+    }
+  }
+}
